Rename yaml-editor-theme to codemirror-theme for the shared surface
